+++ Assets/Constants.cs
@@ -4,5 +4,5 @@ public static class Constants
     public const ushort GroupArea = GroupSize * GroupSize;
     public const byte BitFieldSize = 8;
 
-    public const float PositionMultiplier = GroupSize * (float)BitFieldSize;
+    public const int PositionMultiplier = GroupSize * BitFieldSize;
 }
